Update Simple C# Calculator
Added a prompt  to use the calculator again
Modification 1: Adding a bool variable repeat and setting it to true
This modification adds a boolean variable repeat and initializes it to true. This variable will be used to control the while loop that we'll add later. The idea is that as long as repeat is true, the calculator will continue to run.

Modification 2: Wrapping the calculator logic in a while loop
This modification wraps the entire calculator logic inside a while loop. The loop will continue to run as long as the condition repeat is true. This means that the calculator will continue to prompt the user for input and perform calculations until the user decides to exit.

Modification 3: Adding a check for the user's response to repeat the calculation
This modification adds a check for the user's response to repeat the calculation. Here's what happens:

We prompt the user to input whether they want to use the calculator again.
We read the user's response using Console.ReadLine().
If the user inputs "Y", we use the continue statement to skip to the next iteration of the loop. This means that the calculator will repeat the calculation.
If the user inputs "N", we set repeat to false, which will exit the loop. We also display a farewell message to the user.
If the user inputs anything else, we display an error message and continue to the next iteration of the loop.